+++ src/snowflake/cli/_plugins/spcs/image_repository/commands.py
@@ -15,26 +15,32 @@
 from __future__ import annotations
 
 import json
-from typing import Optional
+from typing import Dict, Optional
 
 import requests
 import typer
-from click import ClickException
+from click import ClickException, UsageError
 from snowflake.cli._plugins.object.command_aliases import (
     add_object_command_aliases,
     scope_option,
 )
 from snowflake.cli._plugins.spcs.image_registry.manager import RegistryManager
+from snowflake.cli._plugins.spcs.image_repository.image_repository_entity_model import (
+    ImageRepositoryEntityModel,
+)
 from snowflake.cli._plugins.spcs.image_repository.manager import ImageRepositoryManager
+from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.commands.flags import (
     IfNotExistsOption,
     ReplaceOption,
+    entity_argument,
     identifier_argument,
     like_option,
 )
 from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
 from snowflake.cli.api.console import cli_console
 from snowflake.cli.api.constants import ObjectType
+from snowflake.cli.api.exceptions import NoProjectDefinitionError
 from snowflake.cli.api.identifiers import FQN
 from snowflake.cli.api.output.types import (
     CollectionResult,
@@ -92,6 +98,46 @@ def create(
             name=name.identifier, replace=replace, if_not_exists=if_not_exists
         )
     )
+
+
+@app.command(requires_connection=True)
+def deploy(
+    replace: bool = ReplaceOption(
+        help="Replace the image-repository if it already exists."
+    ),
+    entity_id: str = entity_argument("image-repository"),
+    **options,
+):
+    """
+    Deploys the image repository from snowflake.yml.
+    """
+    cli_context = get_cli_context()
+    pd = cli_context.project_definition
+    image_repositories: Dict[str, ImageRepositoryEntityModel] = pd.get_entities_by_type(
+        entity_type="image-repository"
+    )
+
+    if not image_repositories:
+        raise NoProjectDefinitionError(
+            project_type="image repository", project_root=cli_context.project_root
+        )
+
+    if entity_id and entity_id not in image_repositories:
+        raise UsageError(f"No '{entity_id}' entity in project definition file.")
+    elif len(image_repositories.keys()) == 1:
+        entity_id = list(image_repositories.keys())[0]
+
+    if entity_id is None:
+        raise UsageError(
+            "Multiple image repositories found. Please provide entity id for the operation."
+        )
+
+    SingleQueryResult(
+        ImageRepositoryManager().deploy(
+            image_repository=image_repositories[entity_id], replace=replace
+        )
+    )
+    return MessageResult(f"Image repository '{entity_id}' successfully deployed.")
 
 
 @app.command("list-images", requires_connection=True)